Did I tell you how much I hate Swing?
The layout managers are especially gruesome.

Here's the UI for adding multipliers. I had to totally change the layout method to get it to look ok. I think that it will give us more flexibility in the future. Feel free of course to change any of it. It could be a lot better.