Fixed ClipLauncher so that the bottom row always is used to stop all clips on that column.
This makes it so that you can launch clips that are set to record at launch for resampling etc...
Otherwise it would not be possible to stop the.

This means now that each view launches a set of clips 7 tracks by 7 clip slots in size with bottom row being for stop.